2020.11.15 file upload complete

+++ dealight_store/src/main/java/com/dealight/domain/StoreLocVO.java
@@ -21,8 +21,8 @@ public class StoreLocVO {
     private String addr;
 
     // 위도 
-    private int lt;
+    private double lt;
 
     // 경도 
-    private int lo;
+    private double lo;
 }
